+++ config/routes.rb
@@ -3,4 +3,9 @@ Rails.application.routes.draw do
 
   # Defines the root path route ("/")
   # root "articles#index"
+
+  namespace :api do
+    resources :grading_jobs
+  end
+
 end
